remove vault fees, dup count
